🗃️ Committing everything that changed 🗃️

+++ templates/scripts/other/docker-entrypoint
@@ -212,9 +212,9 @@ if [ -n "$DEFAULT_TEMPLATE_DIR" ]; then
         create_template_name="$(basename "$create_config_template")"
         if [ -d "$create_config_template" ]; then
           mkdir -p "/config/$create_template_name/"
-          cp -Rf "$create_config_template/." "/config/$create_template_name/" 2>/dev/null
+          [ -e "/config/$create_template_name" ] || cp -Rf "$create_config_template/." "/config/$create_template_name/" 2>/dev/null
         elif [ -e "$create_config_template" ]; then
-          cp -Rf "$create_config_template" "/config/$create_template_name" 2>/dev/null
+          [ -e "/config/$create_template_name" ] || cp -Rf "$create_config_template" "/config/$create_template_name" 2>/dev/null
         fi
       fi
     done
@@ -231,9 +231,9 @@ if [ -n "$DEFAULT_CONF_DIR" ]; then
       if [ -n "$create_config_template" ]; then
         if [ -d "$create_config_template" ]; then
           mkdir -p "/config/$create_config_name"
-          cp -Rf "$create_config_template/." "/config/$create_config_name/" 2>/dev/null
+          [ -e "/config/$create_config_name" ] || cp -Rf "$create_config_template/." "/config/$create_config_name/" 2>/dev/null
         elif [ -e "$create_config_template" ]; then
-          cp -Rf "$create_config_template" "/config/$create_config_name" 2>/dev/null
+          [ -e "/config/$create_config_name" ] || cp -Rf "$create_config_template" "/config/$create_config_name" 2>/dev/null
         fi
       fi
     done
@@ -250,9 +250,9 @@ if [ -d "/data" ]; then
       if [ -n "$create_data_template" ]; then
         if [ -d "$create_data_template" ]; then
           mkdir -p "/data/$create_data_name"
-          cp -Rf "$create_data_template/." "/data/$create_data_name/" 2>/dev/null
+          [ -e "/data/$create_data_name" ] || cp -Rf "$create_data_template/." "/data/$create_data_name/" 2>/dev/null
         elif [ -e "$create_data_template" ]; then
-          cp -Rf "$create_data_template" "/data/$create_data_name" 2>/dev/null
+          [ -e "/data/$create_data_name" ] || cp -Rf "$create_data_template" "/data/$create_data_name" 2>/dev/null
         fi
       fi
     done
@@ -280,13 +280,17 @@ if [ -d "/config" ]; then
 fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 # Copy html files
-if [ "$DATA_DIR_INITIALIZED" = "false" ] && [ -d "$DEFAULT_DATA_DIR/data/htdocs" ]; then
-  cp -Rf "$DEFAULT_DATA_DIR/data/htdocs/." "$WWW_ROOT_DIR/" 2>/dev/null
+if [ "$DATA_DIR_INITIALIZED" = "false" ] && [ -n "$WWW_ROOT_DIR" ]; then
+  if [ -d "$DEFAULT_DATA_DIR/data/htdocs" ]; then
+    [ "$(ls -A "$WWW_ROOT_DIR/" 2>/dev/null | wc -l)" -ne 0 ] || cp -Rf "$DEFAULT_DATA_DIR/data/htdocs/." "$WWW_ROOT_DIR/" 2>/dev/null
+  fi
 fi
-if [ -d "$DEFAULT_DATA_DIR/htdocs/www" ] && [ ! -d "$WWW_ROOT_DIR" ]; then
-  mkdir -p "$WWW_ROOT_DIR"
-  cp -Rf "$DEFAULT_DATA_DIR/htdocs/www/" "$WWW_ROOT_DIR"
-  [ -f "$WWW_ROOT_DIR/htdocs/www/server-health" ] || echo "OK" >"$WWW_ROOT_DIR/htdocs/www/server-health"
+if [ -n "$WWW_ROOT_DIR" ]; then
+  if [ -d "$DEFAULT_DATA_DIR/htdocs/www" ] && [ ! -d "$WWW_ROOT_DIR" ]; then
+    mkdir -p "$WWW_ROOT_DIR"
+    cp -Rf "$DEFAULT_DATA_DIR/htdocs/www/" "$WWW_ROOT_DIR"
+    [ -f "$WWW_ROOT_DIR/htdocs/www/server-health" ] || echo "OK" >"$WWW_ROOT_DIR/htdocs/www/server-health"
+  fi
 fi
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 [ -d "$SSL_DIR" ] || mkdir -p "$SSL_DIR"
